Move `used_number_of_scans` to be an attribute of `NMRSpectroscopy`
to make cardinality 1 stick when also declaring it a subslot of `has_quantitative_attribute`. Otherwise the cardinality from `has_quantitative_attribute` will be used. This has the effect  that in the docs page of  `used_number_of_scans` cardinality and range will be wrongly taken from `has_quantitative_attribute` but on the `NMRSpectroscopy` page all is as it should be.